frontend-utils: Move PlayingContent from Desktop
